Move createProvider to hierarchy-builder/Utils.ts
